[Chore] 3차 배포

+++ src/main/java/com/nhnacademy/heukbaekbookshop/book/dto/response/book/BookResponse.java
@@ -3,7 +3,15 @@ package com.nhnacademy.heukbaekbookshop.book.dto.response.book;
 import com.nhnacademy.heukbaekbookshop.contributor.dto.response.ContributorSummaryResponse;
 import com.nhnacademy.heukbaekbookshop.contributor.dto.response.PublisherSummaryResponse;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record BookResponse(Long id, String title, String publishedAt, String salePrice, double discountRate, String thumbnailUrl, List<ContributorSummaryResponse> contributors, PublisherSummaryResponse publisher) {
+public record BookResponse(Long id,
+                           String title,
+                           String publishedAt,
+                           String salePrice,
+                           BigDecimal discountRate,
+                           String thumbnailUrl,
+                           List<ContributorSummaryResponse> contributors,
+                           PublisherSummaryResponse publisher) {
 }
